Fixed event dispatching inconsistencies on some models leading to incomplete API.
Events are now properly triggered on all models.

Added event support to com_config save actions.

Improved JModelAdmin.

+++ administrator/components/com_menus/models/menu.php
@@ -169,8 +169,12 @@ class MenusModelMenu extends JModelForm
 	 */
 	public function save($data)
 	{
+		$dispatcher = JDispatcher::getInstance();
 		$id	= (!empty($data['id'])) ? $data['id'] : (int)$this->getState('menu.id');
 		$isNew	= true;
+
+		// Include the plugins for the on save events.
+		JPluginHelper::importPlugin('content');
 
 		// Get a row instance.
 		$table = $this->getTable();
@@ -193,11 +197,22 @@ class MenusModelMenu extends JModelForm
 			return false;
 		}
 
+		// Trigger the before event.
+		$result = $dispatcher->trigger('onContentBeforeSave', array($this->_context, &$table, $isNew));
+		if (in_array(false, $result, true))
+		{
+			$this->setError($table->getError());
+			return false;
+		}
+
 		// Store the data.
 		if (!$table->store()) {
 			$this->setError($table->getError());
 			return false;
 		}
+
+		// Trigger the after save event.
+		$dispatcher->trigger('onContentAfterSave', array($this->_context, &$table, $isNew));
 
 		$this->setState('menu.id', $table->id);
 
@@ -215,6 +230,8 @@ class MenusModelMenu extends JModelForm
 	 */
 	public function delete($itemIds)
 	{
+		$dispatcher = JDispatcher::getInstance();
+
 		// Sanitize the ids.
 		$itemIds = (array) $itemIds;
 		JArrayHelper::toInteger($itemIds);
@@ -222,12 +239,34 @@ class MenusModelMenu extends JModelForm
 		// Get a group row instance.
 		$table = $this->getTable();
 
+		// Include the plugins for the on delete events.
+		JPluginHelper::importPlugin('content');
+
 		// Iterate the items to delete each one.
 		foreach ($itemIds as $itemId) {
-			// TODO: Delete the menu associations - Menu items and Modules
 
-			if (!$table->delete($itemId))
-			{
+			if ($table->load($itemId)) {
+
+				// Trigger the before delete event.
+				$result = $dispatcher->trigger('onContentBeforeDelete', array($this->_context, $table));
+				if (in_array(false, $result, true))
+				{
+					$this->setError($table->getError());
+					return false;
+				}
+
+				// TODO: Delete the menu associations - Menu items and Modules
+
+				if (!$table->delete($itemId))
+				{
+					$this->setError($table->getError());
+					return false;
+				}
+
+				// Trigger the after delete event.
+				$dispatcher->trigger('onContentAfterDelete', array($this->_context, $table));
+
+			} else {
 				$this->setError($table->getError());
 				return false;
 			}
